requirements.txt
absl-py==0.6.1
astor==0.7.1
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0.post1
bleach==3.0.2
certifi==2018.10.15
configparser==3.5.0
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.0
grpcio==1.16.0
h5py==2.8.0
ipaddress==1.0.22
ipykernel==4.10.0
ipython==5.8.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Lasagne==0.2.dev1
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==1.5.1
mistune==0.8.4
mock==2.0.0
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.7.0
numpy==1.15.4
opencv-python==3.4.3.18
pandas==0.23.4
pandocfilters==1.4.2
pathlib2==2.3.2
pbr==5.1.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==3.0.0
prometheus-client==0.4.2
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pycairo==1.13.3
Pygments==2.2.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.7
pyzmq==17.1.2
qtconsole==4.4.2
scandir==1.9.0
scikit-learn==0.20.0
scipy==0.18.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
sklearn==0.0
subprocess32==3.5.3
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==0.9.0
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
xgboost==0.81
